+++ app/scripts/app.js
@@ -21,6 +21,7 @@ import { extendedTokenComponent } from "./components/extended/token"
 import { extendedStructTokenComponent } from "./components/extended/struct_token"
 import { extendedCQPValueComponent } from "./components/extended/cqp_value"
 import { kwicComponent } from "./components/kwic"
+import { trendDiagramComponent } from "./components/trend_diagram"
 import * as treeUtil from "./components/corpus_chooser/util"
 import statemachine from "@/statemachine"
 
@@ -80,6 +81,7 @@ korpApp.component("extendedToken", extendedTokenComponent)
 korpApp.component("extendedStructToken", extendedStructTokenComponent)
 korpApp.component("extendedCqpValue", extendedCQPValueComponent)
 korpApp.component("kwic", kwicComponent)
+korpApp.component("trendDiagram", trendDiagramComponent)
 
 // load all custom components
 let customComponents = {}
